merge dal main

+++ ItineraJava/src/main/java/com/unipi/ItineraJava/controller/CommunityController.java
@@ -5,6 +5,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
+import com.unipi.ItineraJava.service.auth.JwtTokenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -27,22 +28,29 @@ class CommunityController {
     @Autowired
     private CommunityService communityService;
     @Autowired
+
     private CommunityNeo4jRepository communityNeo4jRepository;
     @Autowired
+
     private CommunityRepository mongoCommunityRepository;
+    @Autowired
+    private User user;
 
 
+    // http://localhost:8080/Community
+    // returns all communities with details
     @GetMapping
-    public ResponseEntity<List<MongoCommunity>> getAllCommunityNames() {
+    public ResponseEntity<List<MongoCommunity>> getAllCommunity() {
         List<MongoCommunity> communities = mongoCommunityRepository.findAll();
         System.out.println(communities);
         return ResponseEntity.ok(communities);
     }
 
-
+    // http://localhost:8080/Community/678e41769d6b117cd029652e
+    // returns the {id} community with all his data
     @GetMapping("/{id}")
     public Optional<MongoCommunity> getCommunityById(@PathVariable String id) {
-        return communityService.findById(id);
+        return ResponseEntity.ok(communityService.findById(id)).getBody();
     }
 
     @GetMapping("/details")
@@ -58,7 +66,32 @@ class CommunityController {
         return null;
     }
 
-    // http://localhost:8080/Community with body
+    // adds a post inside a community
+    @PutMapping("community/{name}")
+    public ResponseEntity<String> updateCommunity(@RequestHeader("Authorization") String token,
+                                                  @RequestBody String text,
+                                                  @RequestParam String name)
+    {
+        try{
+            String username = JwtTokenProvider.getUsernameFromToken(token);
+            if (username == null)
+                return ResponseEntity.status(400).body("Invalid token");
+            if (text == null)
+                return ResponseEntity.status(400).body("Invalid text");
+            if(communityService.existsCommunity(name)) {
+                return communityService.updateCommunity(username, text, name);
+            }else
+            {
+                return ResponseEntity.status(400).body("Invalid Community name");
+            }
+        }catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
+        }
+    }
+
+
+    // http://localhost:8080/Community
+    //creates a community checking if the admin is sending the request
     @PostMapping
     public ResponseEntity<String> createCommunity(
             @RequestHeader("Authorization") String token,
@@ -79,10 +112,22 @@ class CommunityController {
             return ResponseEntity.status(500).body("Error creating community: " + e.getMessage());
         }
     }
-
+    // http://localhost:8080/Community/Viareggio
+    // deletes a community
     @DeleteMapping("/{name}")
-    public void deleteCommunity(@PathVariable String name) {
-        communityService.deleteByName(name);
+    public ResponseEntity<String> deleteCommunity(@RequestHeader("Authorization") String token,
+                                                  @PathVariable String name) {
+        try {
+            if (User.isAdmin(token)) {
+                communityService.deleteByName(name);
+            } else {
+                return ResponseEntity.status(400).body("User not authenticated as Admin");
+            }
+            return ResponseEntity.ok("Community deleted successfully");
+        }catch(Exception e){
+            return ResponseEntity.status(500).body("Error deleting community: " + e.getMessage());
+        }
+
     }
